:sparkles: SVGKeyboard コンポーネントで Keyboard コンポーネントを置き換えた
