--- src/components/blackKey/BlackKey.css
@@ -1,17 +1,0 @@
-.blackKey {
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 35px;
-  width: 50px;
-  height: 125px;
-  background-color: rgb(58, 58, 58);
-  z-index: 1;
-  border: none;
-  appearance: none;
-  cursor: pointer;
-}
-
-.blackKey:hover {
-  background-color: black;
-}
