+++ src/components/index.ts
@@ -2,6 +2,5 @@ export * from './analyzer';
 export * from './detuneKnob';
 export * from './gainKnob';
 export * from './keyboard';
-export * from './svgKeyboard';
 export * from './unisonKnob';
 export * from './waveformSelectbox';
