--- src/components/key/Key.css
@@ -1,5 +1,0 @@
-.key {
-  position: relative;
-  width: 60px;
-  height: 250px;
-}
