--- src/components/key/index.tsx
@@ -1,5 +1,0 @@
-import styles from './Key.css';
-
-export const Key: React.FC = ({ children }) => (
-  <div className={styles.key}>{children}</div>
-);
