--- src/components/keyboard/Keyboard.css
@@ -1,3 +1,0 @@
-.keyboard {
-  display: flex;
-}
