--- src/components/whiteKey/WhiteKey.css
@@ -1,24 +1,0 @@
-.whiteKey {
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 60px;
-  height: 230px;
-  background-color: white;
-  border-top: 1px solid black;
-  border-left: none;
-  border-right: 1px solid black;
-  border-bottom: 1px solid black;
-  z-index: 0;
-  appearance: none;
-  cursor: pointer;
-}
-
-.leftEnd {
-  border-left: 1px solid black;
-}
-
-.whiteKey:hover {
-  background-color: rgb(224, 224, 224);
-}
